Factor out PAG logic to metrics package
